this instead

+++ travis_setup.sh
@@ -13,8 +13,12 @@ cd ../../..;
 
 
 echo 'Make .modules';
-mkdir ~/.modules;
-mkdir ~/.modules/modulefiles;
+if [ ! ~/.modules ]; then
+    mkdir ~/.modules;
+fi
+if [ ! ~/.modules ]; then
+    mkdir ~/.modules/modulefiles;
+fi
 sudo chown -R travis:travis ~/.modules
 
 # Create the necessary conda environments
@@ -27,27 +31,31 @@ if [ ! -d ~/.modules/mmeds-stable ]; then
     Rscript setup.R &> /dev/null;
     source deactivate;
 fi
+
 if [ ! -d ~/.modules/qiime1 ]; then
     echo "Create qiime1 environment";
     conda create python=2.7 qiime matplotlib=1.4.3 mock nose -c bioconda --yes --quiet --copy -p ~/.modules/qiime1 &>/dev/null;
 fi
-rm -rf ~/.modules/qiime2
+
+rm -rf ~/.modules/qiime2;
 if [ ! -d ~/.modules/qiime2 ]; then
     echo "Create qiime2 environment"
     wget https://data.qiime2.org/distro/core/qiime2-2019.1-py36-linux-conda.yml -O ~/qiime2.yml --quiet;
-    conda env create --file ~/qiime2.yml --quiet -p ~/.modules/qiime2 &>/dev/null;
+    conda env create --file ~/qiime2.yml --quiet -p ~/.modules/qiime2;
 fi
 
-# Configure and install environment modules
-mkdir ~/.local
-echo "Install environment-modules";
-wget https://sourceforge.net/projects/modules/files/Modules/modules-4.2.1/modules-4.2.1.tar.gz -O modules-4.2.1.tar.gz &>/dev/null;
-tar -zxf modules-4.2.1.tar.gz;
-cd modules-4.2.1;
-./configure --prefix="${HOME}/.local" --modulefilesdir="${HOME}/.modules/modulefiles" &>/dev/null;
-make &>/dev/null;
-sudo make install &>/dev/null;
-cd $REPO_DIR;
+if [ ! -f ~/.local/init/profile.sh ]; then
+    # Configure and install environment modules
+    mkdir ~/.local
+    echo "Install environment-modules";
+    wget https://sourceforge.net/projects/modules/files/Modules/modules-4.2.1/modules-4.2.1.tar.gz -O modules-4.2.1.tar.gz &>/dev/null;
+    tar -zxf modules-4.2.1.tar.gz;
+    cd modules-4.2.1;
+    ./configure --prefix="${HOME}/.local" --modulefilesdir="${HOME}/.modules/modulefiles" &>/dev/null;
+    make &>/dev/null;
+    sudo make install &>/dev/null;
+    cd $REPO_DIR;
+fi
 
 # Make sure module will work
 export PATH="$HOME/.local/bin:$PATH"
